split linkchecker to pkg and utils
