+++ .gitignore
@@ -1,4 +1,4 @@
-linkchecker
+bin/
 .idea
 coverage.out
 .vscode
