+++ Taskfile.yml
@@ -3,7 +3,7 @@ version: '3'
 tasks:
   build:
     cmds:
-      - go build -o linkchecker cmd/linkchecker/main.go
+      - go build -o bin/linkchecker cmd/linkchecker/main.go
     silent: true
 
   test:
